requirements.txt
pandas>=2.2
requests>=2.31
sqlalchemy>=2.0
python-dotenv>=1.0
matplotlib>=3.9
jinja2>=3.1
openai>=1.51
langchain>=0.3
langgraph>=0.2
pdfkit>=1.0
seaborn>=0.13
pytest>=8.2,<9
xhtml2pdf>=0.2.15
reportlab>=4.2